chapter2-extracting-data-from-pdfs/extracting-pdf-tables/requirements.txt
tabula-py
camelot-py[cv]
pdfplumber
pandas